WebGL support

+++ include/Galaxy/Renderer/mesh.hpp
@@ -1,6 +1,7 @@
 #pragma once
 #include "Galaxy/Math/vector.hpp"
 #include <vector>
+#include <Galaxy/OS/defines.hpp>
 
 class VertexArray;
 class VertexBuffer;
@@ -17,6 +18,11 @@ struct Vertex
     Vertex(float x, float y, float z) : pos(x, y, z) { }
     Vertex(Vector3 pos, Vector2 texCoord = Vector2(), Vector3 normal = Vector3()) : pos(pos), texCoord(texCoord), normal(normal) { }
 };
+//#if OS_WEB
+//typedef unsigned short TriangleIndex;
+//#else
+//typedef unsigned int TriangleIndex;
+//#endif
 class Mesh
 {
     private:
@@ -28,7 +34,7 @@ class Mesh
 
     public:
         std::vector<Vertex> verts;
-        std::vector<unsigned short> tris;
+        std::vector<unsigned int> tris;
         Vector3 aabbMin, aabbMax;
 
         void initialize_mesh();
@@ -37,7 +43,7 @@ class Mesh
         void calculate_normals();
 
         Mesh();
-        Mesh(std::vector<Vertex> verts, std::vector<unsigned short> tris);
+        Mesh(std::vector<Vertex> verts, std::vector<unsigned int> tris);
         ~Mesh();
 
         Mesh(const Mesh&) = delete;
